chore(*): Move DataTrail files to the AppDataTrail folder
