+++ src/App/AppContext.tsx
@@ -1,18 +1,18 @@
 import { createContext, useContext } from 'react';
 
-import { type DataTrail } from 'DataTrail';
+import { type DataTrail } from 'AppDataTrail/DataTrail';
 import { newMetricsTrail } from 'utils';
 
-interface MetricsAppContextState {
+interface AppContextState {
   trail: DataTrail;
 }
 
 export const defaultTrail = newMetricsTrail();
 
-export const MetricsAppContext = createContext<MetricsAppContextState>({
+export const AppContext = createContext<AppContextState>({
   trail: defaultTrail,
 });
 
 export function useMetricsAppContext() {
-  return useContext(MetricsAppContext);
+  return useContext(AppContext);
 }
