+++ src/App/Routes.tsx
@@ -2,9 +2,9 @@ import React, { lazy } from 'react';
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 
 import { ROUTES } from '../constants';
-import { useMetricsAppContext } from './MetricsAppContext';
+import { useMetricsAppContext } from './AppContext';
 
-export const Wingman = lazy(() => import('../pages/TrailWingman'));
+export const Trail = lazy(() => import('./Trail'));
 
 // For /trail links, redirect to /drilldown with the same search params
 const TrailRedirect = () => {
@@ -17,7 +17,7 @@ export const AppRoutes = () => {
 
   return (
     <Routes>
-      <Route path={ROUTES.Drilldown} element={<Wingman trail={trail} />} />
+      <Route path={ROUTES.Drilldown} element={<Trail trail={trail} />} />
       <Route path={ROUTES.Trail} element={<TrailRedirect />} />
       {/* catch-all route */}
       <Route path="*" element={<Navigate to={ROUTES.Drilldown} replace />} />
