+++ src/AppDataTrail/header/PluginInfo/PluginLogo.tsx
